mm/filemap: return only head pages from find_get_entries
All callers now expect head (and base) pages, and can handle multiple head
pages in a single batch, so make find_get_entries() behave that way.  Also
take the opportunity to make it use the pagevec infrastructure instead of
open-coding how pvecs behave.  This has the side-effect of being able to
append to a pagevec with existing contents, although we don't make use of
that functionality anywhere yet.

Link: https://lkml.kernel.org/r/20201112212641.27837-17-willy@infradead.org
Signed-off-by: Matthew Wilcox (Oracle) <willy@infradead.org>
Reviewed-by: Jan Kara <jack@suse.cz>
Reviewed-by: William Kucharski <william.kucharski@oracle.com>
Reviewed-by: Christoph Hellwig <hch@lst.de>
Cc: Dave Chinner <dchinner@redhat.com>
Cc: Hugh Dickins <hughd@google.com>
Cc: Johannes Weiner <hannes@cmpxchg.org>
Cc: Kirill A. Shutemov <kirill.shutemov@linux.intel.com>
Cc: Yang Shi <yang.shi@linux.alibaba.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>